requirements.txt
pyserial
flask
request
